requirements.txt
apache-airflow
faker
pandas
apache-airflow-providers-postgres
psycopg2-binary
# psycopg2